// src/app/models/prisoner.ts
/**
 * Created by smtandabuzo on 2019/08/05.
 */
export class Prisoner {
    nationalID: number;
    fileNumber: number;
    dateOfBirth: Date;
    sentence: string;
    location: string;
    fullName: string;
    address: string;
    prison: string;
    gender: string;
    education: string;
    status: string;
    offence: string;
    dateIn: Date;
    public prisoners: any = [];
}